Refactor log types and checker error in preparation for output parsing
